Fix comments + workflow run
