+++ packages/twenty-front/src/modules/workflow/workflow-steps/components/WorkflowStepDetail.tsx
@@ -17,8 +17,8 @@ import { WorkflowEditActionHttpRequest } from '@/workflow/workflow-steps/workflo
 import { WorkflowEditActionIterator } from '@/workflow/workflow-steps/workflow-actions/iterator-action/WorkflowEditActionIterator';
 import { WorkflowEditTriggerCronForm } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerCronForm';
 import { WorkflowEditTriggerDatabaseEventForm } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerDatabaseEventForm';
-import { WorkflowEditTriggerManualForm } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerManualForm';
-import { WorkflowEditTriggerManualV2 } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerManualV2';
+import { WorkflowEditTriggerManual } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerManual';
+import { WorkflowEditTriggerManualDeprecated } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerManualDeprecated';
 import { WorkflowEditTriggerWebhookForm } from '@/workflow/workflow-trigger/components/WorkflowEditTriggerWebhookForm';
 import { useIsFeatureEnabled } from '@/workspace/hooks/useIsFeatureEnabled';
 import { assertUnreachable, isDefined } from 'twenty-shared/utils';
@@ -76,7 +76,7 @@ export const WorkflowStepDetail = ({
         case 'MANUAL': {
           if (isIteratorEnabled) {
             return (
-              <WorkflowEditTriggerManualV2
+              <WorkflowEditTriggerManual
                 key={stepId}
                 trigger={stepDefinition.definition}
                 triggerOptions={props}
@@ -85,7 +85,7 @@ export const WorkflowStepDetail = ({
           }
 
           return (
-            <WorkflowEditTriggerManualForm
+            <WorkflowEditTriggerManualDeprecated
               key={stepId}
               trigger={stepDefinition.definition}
               triggerOptions={props}
